review recommendation fix
1. remove python from .tool-versions
2. move DownloadArchive's rc validations to withDefauts()
3. refactor client.go to several other files (archive.go, hooks.go, retry.go, status.go)
