+++ .tool-versions
@@ -7,4 +7,4 @@ kubeconform 0.7.0
 kustomize 5.8.0
 mockery 3.3.4
 tilt 0.33.2
-python 3.11.10
+
